Remove local rollup rule validation
Instead of doing validation locally we will start to fully rely on
dry run validation that we do for all resources that support it.

The rationale behind this is that changes/additions to the APIs
may render a local check obsolete and require a new version of
the Terraform provider when it is not necessary. This has been
seen recently as the result of additionally functionality added
to rollup rules