Block duplicate routes with same severity [sc-77636] (#22)
The TF schema models routes as a list, where each element specifies the
severity. This allows for multiple routes with the same severity, which
is unsupported, and leads to ignored notifiers currently.

Instead of ignoring silently, or trying to merge the routes (which isn't
always possible, e.g., with different repeat-intervals), block duplicate
routes. The route contains a list of notifiers, so users can merge the
lists, which is simpler in the config, and in the implementation.