Support migration from storage_policies to interval [sc-81013]
[sc-81013]

Currently, migrating from storage_policies to interval is not possible
as the provider receives both fields set, even if storage_policies is no
longer set. This seems like an issue with the plugin sdk, as it's
passing the last returned server value, even if the user deletes the
config.

However, we can workaround this by:
 * Ensuring the config doesn't allow both to be set by using
   `ConflictsWith`, which only triggers if the config has both set.
 * Clearing storage_policies if it's set alongside interval, since
   we know that both cannot be set due to the above validation.