registry: Enable DryRun by default, make it opt-out
DryRun is enabled for almost all resources, and it's easy to forget to
enable for new resources, so flip the default so it's enabled by default
and requires opting out.

The only resource without DryRun is TraceMetricsRule -- it's likely this
was missed due to the opt-in nature, but it's not enabled in this PR to
keep this change a pure refactor with no behavioural changes.