Simplify enums: replace legacy type with legacyAlias string
Legacy/V1 was used when migrating the provider to use the public API.
All resources now use the public API, so we can simplify the enum
support to only support V1 types for Swagger, and keep legacy aliases as
strings around for backwards compatibility.

As part of this, ensure that legacyAliases are only specified if they
are unique, to minimize noise.