fix release version check
In the refactoring that was done for open-sourcing the Terraform
provider, the GIT_VERSION environment variable was moved out of
Makefile and into release scripts. This caused a few make
targets to output the wrong thing, but also caused releases
to fail release branch validation. This required that releases
using `SKIP_RELEASE_BRANCH_VALIDATION=true` which leave potential
for human error when making a release.

This PR moves GIT_VERSION back into Makefile and propagates it
along through the release scripts