[tracing] Remove deprecated fields and propagate errors when converting a TraceSearchFilter to model (#44)
The three fields: `TraceSpanFilter.Tags`, `Duration.MaxSeconds`,
`Duration.MinSeconds` currently are not accepted by the terraform
provider and return an error. Due to the error itself being dropped,
this results in the whole TraceSearchFilter being set to nil, and the
backend API being called with an empty search filter. In this PR we
remove these invalid fields so that error messages will correctly
indicate what the issue is, as well as routing errors up the chain so
that errors from validating fields such as `TraceSpanFilter.MatchType`
will be propagated upwards and return a useful error message.