log_allocation_config: Fix dataset reference
The TF ecosystem uses `_id` fields for references to other objects.
intschema also relies on the `_id` suffix to generate a field as a
`tfid.ID`, allowing tests to use references.

We also need to handle dry-run validation when using a reference
as the ID isn't known at plan time if it hasn't been created, so
we replace it with a dummy value.