tfschema: Fix rollup/mapping rule interval diff after apply [sc-111204]
[sc-111204]

When neither interval or storage_policy is specified, a server-side
default is returned. This causes a diff-after-apply, so mark the field
as computed.

Note: When using a computed field, if the user sets an explicitly value,
say `interval=60s`, and then removes that field, Terraform cannot
differentiate whether the value in the state is a server-side default or
previously set, and will continue to use the existing value as-is.