Support unknown dynamic values for monitor.query [sc-114507]
[sc-114507]

If a value is unknown at plan time, terraform sends an empty value. This
can happen if for example, a monitor query contains an interpolated ID,
but the ID won't be known till apply time. When this happens, the entire
query is sent as an empty string at plan.

We can check whether the raw config has a value for this field. For
top-level fields, this is simple as we can use the field name as-is, but
monitor queries are in nested structs, which are modeled as a list hence
an additional `Index(0)` lookup.

Update the setUnknownReferences logic to handle unknown primitives such
as monitor queries, and update monitors to default an unknown dynamic
query to a dry-run default value.