Don't convert empty monitor conditions back to API model
Our Pulumi provider is built using Pulumi's Terraform
bridge library. We ran into a bug specific to Pulumi where
updating the conditions of a monitor after it had been
created caused validation to fail before we even got
calling the API. The issue was trivially reproducible
in Pulumi, but non-reproducible with the same
configuration in Terraform. Upon debugging I found
that the code that converts monitor condition state
representation to the API model format was getting
an empty condition. This shouldn't happen if the
monitor was successfully applied in the past and
stored to state.

The fix I made was to ignore condition args that
are fully empty (no op, no value, no severity, etc)

I believe that this is a result of the
difference in how Terraform and Pulumi persist state.

sc-118429