Fix recording-rule dry-run when bucket/group are set [sc-125838]
When a recording rule has `execution_group` and a `bucket_slug` set,
they're expected to match. However at dry-run time, the `bucket_slug`
may not be known yet and is set to a dummy value which won't match
`execution_group` and fail to validate.

Update validation to use the `execution_group` as the `bucket_slug`
if it's unknown.

As part of this, use a clearer value for unknown dry-run references.