unstable chronosphere_consumption_config resource (#132)
Implements new `chronosphere_consumption_config` resource
against unstable API.

This is our first resource which has a recursive structure
(`partitions`), which Terraform does not support. The
approach taken is to:
1. generate a dynamically nested resource in tfschema
   against a max depth (see makeRecursiveResource).
2. in generateintschema, we allowlist the field as recursive,
   where we stop `newStructDef` recursion early to prevent
   the "real" tfschema from generating N types. Basically,
   instead of this:

```
type Partition struct {
  partition []PartitionPartition
}

type PartitionPartition struct {
  partition []PartitionPartitionPartition
}

...
```

We get this:

```
type Partition struct {
  partition []Partition
}
```


The benefit of this approach is that the intschema<->model
mpaping layer stays simple, while toResourceData and
fromResourceData "just work".

The downside is that dry run breaks, because it ends up
infinitely recursing on the intschema struct while trying
to populate dummy references in `setUnknown`. I've opted
to not enable dry run validation for ConsumptionConfig in
this PR, to keep this PR straightforward. I'll follow up w/
a dry run fix in a separate PR.