Fix `make install-snapshot` for goreleaser v2.11.1 (#139)
When upgrading to goreleaser v2, there were some breaking
changes to the `make install-snapshot` target. The main
one is that goreleaser now includes arch versions (e.g. GOARM64 or
GOAMD64)
in the build directory name:
https://goreleaser.com/customization/builds/go/#why-is-there-a-_v1-suffix-on-amd64-builds

It doesn't seem there's a way to disable this, so
we have to work it into our Makefile.